containers: Fix docker build path and lowercase version parameter
build-dockerimage.sh doesn't live in the src/chromiumos path, and
having mixed-case in the version string passed to the script causes
issues with uploading the finished docker image to the registry.

We return the final registry path to the caller, so lowercasing
won't affect any of the logic on the caller side.

BUG=none
TEST=unittests

Change-Id: I1cee114823a9cc6e92a86028123fc45ad8169fd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3119590
Tested-by: Sean McAllister <smcallis@google.com>
Auto-Submit: Sean McAllister <smcallis@google.com>
Commit-Queue: Sean McAllister <smcallis@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>