cbuildbot_launch: Add option to ignore preload for overlayfs mount
Add option to skip performing a preload if the option is provided, to
allow for testing using overlayfs again.  The intent is to change the
location of the buildroot, along with an overlayfs mount, to utilize
snapshot mounts.  This will speed up the sync process, align with the
standard approach, and support branches in the future.

BUG=b:198442885
TEST=`local tryjob`

Change-Id: Ida23a8668465fe2626a8b31551abcfc7aa5d2645
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3138413
Tested-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>