cbuildbot_launch: Flag control the buildroot cleanup
As we consider moving to overlayfs for legacy builders, we need an
option for testing the buildroot maintenance and cleanup.  This will
flag control the cleanup of the buildroot based on whether we utilize
source-cache.  This should be transparent as the flag is set via the recipe and can be verified in a LED job.

BUG=b:197678613
TEST='local tryjob'

Change-Id: I617e1c0008ecd86832e1e3cd5a1278a534c99c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3140363
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>