ap_firmware_config: add generic.py
Pursiuing multiple goals with this change.
1. Better template.
Using generic config as a template helps fill it in. There were many
fields in this config that weren't clear, and we seen folks struggle
to fill them in. Should be a bit clearer now due to comments and
less fields in general.

2. Better config.
extra_flags functions is a flexible future-proof mechanism to add flags
like --fast and --force to the flashing arguments. Replaces
is_fast_required function.
Removed forcing of futility, because it is default, and folks
overriding defaults shouldn't be stopped from doing so. It also makes
code more confusing.
Added a note that BUILD_CONFIG is not neccesarry -> deprecating it.

3. Improved defaults:
cros ap build does not work for many boards because they didn't fill
in the BUILD_CONFIG. However, BUILD_CONFIG should be now just
'chromeos-bootimage'
We know the programmer and dut_control that could be used in most cases,
which would work in many cases. In other cases it would serve as a good
starting point.

BUG=b:199301343
TEST=manual + CQ

Change-Id: I93ab5cb8f8b3c7415d14fbeb8d28ad661d4e873b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3149333
Tested-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>