Working around gerrit error
Gerrit is giving me trouble with
https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3149333
> Merge conflict detected! Please rebase and retry.
I rebased successfully. Retried and got same error again.

> Applied patch to 'lib/firmware/ap_firmware_config/__template.py' with conflicts.
> error: lib/firmware/ap_firmware_config/generic.py: does not exist in index
> error: cannot read the current contents of 'lib/firmware/ap_firmware_config/generic.py'

In previous CL, I attempted to remove __template.py and replace it with a
symlink to generic.py.
It seems that gerrit tried to follow the link and calculate the
diff against current (non-existent) generic.py

Splitting the CL into 2.
Previous doesn't touch __template.py
This one actually replaces it with a symlink

BUG=None
TEST=Gerrit doesn't reject my patch for no reason

Change-Id: I4cd99a473405b502a6c22bab34fda458c0ef43b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3151277
Tested-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>