LAKITU: Increase retry options for SSH into a VM.
Recently, lakitu-arm64 builds have been failing during the VMTests. This is happening because the VM created for test is not completely booted up in case of ARM.
Due to this, the first test fails due to SSH problem. This CL is increasing the number of tries for SSH when a VM is created.

BUG=b/197269863
TEST=multiple presubmit for lakitu-arm64.
RELEASE_NOTE=None

Change-Id: Iddcc24ab48e475b6300b325a4839003f6f71dd20