chrome_chromeos_lkgm: Use owner rather than author when looking for CLs.
When the gardener uploads a new patchset to resolve a merge conflict,
that causes the 'author' of the CL to change. eg:
https://chromium-review.googlesource.com/c/chromium/src/+/3179800

So querying for 'author' of the CrOS robot account ignores that
CL. But the 'owner' doesn't change. So this switches to query off that.

Also, add some additional logging and add '--debug' to its execution on
the master builders so we can more easily debug it on the bots.

BUG=b:187790931
TEST=unittest

Change-Id: I1e803bcca2d624ecbb1b262534b11795f4d7402d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3179868
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Navil Perez <navil@google.com>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>