chrome_committer: Don't CQ+2 the uploaded CL by default.
As part of b/187790931, we'd like to split the LKGM upload into two
phases: the first phase uploads the CL, then the second phase CQ's it.
By default, chrome_committer.py will CQ+2 the CLs it uploads if you
don't tell it to dry-run it.

In the first phase, we don't want the CL dry-runned or submitted. So
this adds a third option to chrome_committer.py that will do neither.

BUG=b:187790931
TEST=unittest

Change-Id: Ib23b2154277589721d4e3fd99c52d6b8d6fe702c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3191690
Tested-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>