build_stages: Do not unmount tree if cache mount
Do not unmount the tree if the build root is a source cache mount.  This
is legacy code that is meant to work with local directories while source
cache mounts will be disk mounts.

This change consolidates past changes to allow for a smooth cherrypick to subsequent branches.

BUG=b:197678613
TEST='tryjob && local tryjob'

Change-Id: Id2b33abfee141e871d66c30da51e6d804bdec758
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3205726
Reviewed-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>