chrome_chromeos_lkgm: Quote the 'message' field in gerrit queries.
Right now the script is running a query similar to:
https://chromium-review.googlesource.com/q/message:LKGM+14263.0.0+project:chromium/src

When it really should be running:
https://chromium-review.googlesource.com/q/message:%2522LKGM+14263.0.0%2522+project:chromium/src

The lack of quotes means it's only looking for "LKGM" in the commit
message, which returns all the LKGM CLs when we really just want the CL
for the given version.

Adding quotes around the message should fix it.

BUG=b:202448086
TEST=unittest

Change-Id: I0713beab00fabf32a261b89a05ea8bc241d62f9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3211864
Tested-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>