cros lint: Enable seccomp linter.
Because of a code oversight the seccomp linter was not actually
enforcing anything. Fix that.

The linter will lint the entire file so we cannot enable the linter for
all dangerous syscalls quite yet. This would result in most seccomp
policies in Chrome OS throwing a linter error, which we don't want.
Instead, enable the linter for the subset of syscalls not currently
used by any policy.

BUG=b:202190205
TEST=cros lint <a policy with clone()> doesn't return an error.

Cq-Depend: chromium:3221135
Change-Id: I7942b47e6434927f6a6107467bbaf71198c3a198
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3221911
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>