pylint-quotes: shuffle import paths to simplify usage
This used a dedicated subdir to avoid having it implicitly added to
sys.path for users, but since we've cleaned up chromite.third_party
to not do that anymore, we can rename this dir.  This way we don't
have to patch sys.path in the pylint module to import it.

BUG=None
TEST=`cros lint` still works & detects bad quotes

Change-Id: Ic165946d535070fe61a832c7906fe8bd8efec487
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3245315
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>