xz_auto: disable pixz's special tar indexing bits
At the moment, running `xz_auto` on an empty file with `pixz` on `$PATH`
breaks. It seems this related to pixz's tar file autodetection, which
isn't something we really make use of at the moment. Disable it, since
it seems to have no impact on pixz's ability to decompress things in
parallel.

Unfortunately, disabling this on both ends (compression and
decompression) seems like it leads to significant badness if we try to
decompress tar'ed files which have the index (...of which, we have a few
scattered throughout). The simplest solution to this seems to be hacking
around the place where badness exists.

BUG=b:202735786
TEST=Unittests pass

Change-Id: I0ba50ec8094a881d0926307717e9e7020c9d842d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3248615
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>