cros_sdk: simplify exit status handling
The big comment block here about exit behavior with interactive mode
doesn't make much sense as it's currently written.  That's because it
was for code that used to print an error message before exiting.  But
when that logic changed to not print a message, the comment block was
left in place.  Since we want to keep the behavior of passing the exit
code back up, let's simplify by deleting the comment and just passing
the exit status back all the time.

This means now we pass the non-zero exit status back when in interactive
mode, but as the stale comment explains, that situation shouldn't really
matter that much.

This also means the decision to exit or return a value is left to the
main loop rather than this helper function.  That makes more sense, and
will help in follow up commits where we move these into libraries.

BUG=b:191307774
TEST=`cros_sdk` still works

Change-Id: I619ddf65fb10e17df570bffbac5661fbe361a0d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3252647
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>