ap_firmware_config: Add default flags to speed up the flash process
When using servo over CCD, it is really slow to read the fw twice
during the flash process. Add -n for flashrom and --fast for
futility which skips the verify step. Changes are not reverted
even when verification fails leaving users in a broken state
regardless of the verify step so we should skip it.

BUG=b:201789808
TEST=CQ

Change-Id: I9b04b99ac2d44a03e0568fa21c17ad5aa3d1e792
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3254855
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Cindy Lin <xcl@google.com>
Tested-by: Cindy Lin <xcl@google.com>