payload service: remove key parameter
We don't need to receive the key parameter for the images as if we're
getting a signed image the correctly formatted key name will be present
by parsing the target image artifact location.

We're removing this from the recipes side and config as well.

BUG=b:177903628
TEST=./run_tests

Change-Id: Ie74f13761de66a91e1d50e57e25ad3db9e7160a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3291946
Tested-by: George Engelbrecht <engeg@google.com>
Auto-Submit: George Engelbrecht <engeg@google.com>
Commit-Queue: Jack Neus <jackneus@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>