api: fetch C++ lints in toolchain service
This CL adds the logic to the Build API to support fetching lints in C++
for use with the build_linters api.

This CL also implements additional filters on findings to only return
results which were affected by the latest git commit. This works for the
build_linters recipe which is currently the only consumer since the
recipe only applies one CL for testing. This is necessary because it was
observed that there may be 100s of preexisting lints in some
repositories which we do not want sent to gerrit. Gerrit does implement
it's own filters for comments on unchanged lines, but also has a limit
to how many comments will be processed.

BUG=b:187790543
TEST=Tested with generated call_scripts

Change-Id: I60108316421165f74b4085103b8152fe6a9d1e4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3299359
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>