cros workon: Fix error working on host packages
When IsFullManifest does not return true, the ebuild lookup would fail
for SDK packages due to limited BOARD_OVERLAYS, which was expected to be
populated with, effectively, all of the overlays it uses. Instead, do the
lookup using equery which, which does not need the value to do the lookup.

BUG=b:206824602
TEST=run_tests, cros workon --host start ap-firmware-config

Change-Id: I360bef797c073f80a77a393a7e0a76dfa7667f3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3307819
Tested-by: Alex Klein <saklein@chromium.org>
Auto-Submit: Alex Klein <saklein@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>