chrome_chromeos_lkgm: Keep LKGM CLs in WIP until they're ready to land
Theoretically an LKGM CL is only ready to land once both the
master-release and master-full builders are done. We currently
upload the CL when the first is done, then CQ+2 it when the second
builder is done.

But sometimes the gardener will see the LKGM CL and CQ+2 it
themselves too early. So this keeps the CL in WIP mode until it's
ready to land. Hopefully that'll prevent emails from going out.

BUG=none
TEST=unittest

Change-Id: Ia8f9e13008b4bfef8a8ff067d6eadc72e56b307e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3308374
Tested-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>