repo_sync_manifest: Remove BuildRootGitCleanup call
The disk mounts for legacy builds are recreated on every execution
therefore the cleanup is unnecessary and results in a sync on main
taking hours to remove and hours to sync again.  We can assume the repo
is clean to start, as it is a fresh checkout, and can skip the
housekeeping.

BUG=b:207032641
TEST='cros tryjob'

Change-Id: I44dba9995b03e6f29841a2d55b3262ac3f448cc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3307088
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>