chrome_committer: Don't branch off of 'master' when checking out.
Latest LKGM CLs are having rebase errors:
https://chromium-review.googlesource.com/c/chromium/src/+/3306603

Presumably because their parent revision is:
https://chromium.googlesource.com/chromium/src/+/fd8a8914ca0183f0add65ae55f04e287543c7d4a

Which is almost 3 weeks old, and is not coincidentally when
'master' stopped getting updated. So this removes the 'master'
pin and lets the Checkout() and ShallowFetch() methods use the default
of HEAD, which should work fine.

BUG=b:205987120
TEST=unittest

Change-Id: Iec3f9d2be13a06dec7a5765f663ec8ffab7b0163
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3309465
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>