Lacros: add --is-lacros support to cros_chrome_sdk
Currently, Lacros uses the simple chrome cros toolchain that is updated
on a roughly daily basis, however, there has been quite a few unfixable
issues due to version skew that results in release blockers.

The long term solution is for Lacros to switch to use browser toolchain,
but it's not happening soon.

This CL implements a short term workaround to require Lacros uses a
passed in version number for compiling Lacros, but meanwhile, Lacros
will continue to use the version defined in CHROMEOS_LKGM for testing
purposes.

BUG=chromium:1204195
TEST=wrote and ran unit tests

Change-Id: I0559a96c23a0a929195a1d164703129836404f8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3308954
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Tested-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>