call_templates: add get_builder_metadata_input.json
Add a template for this call. We need the model and firmware versions
and metadata to push to goldeneye. Adding this template was just
confirming that the buildapi supports this. We need to make the recipe
changes to actually call this in rubik (maybe other?) builders.

BUG=b:206154671
TEST=ran manually, works

Change-Id: I7b1a084dddebc9de5e8cf9f9cb0490a56d6b29d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3328089
Auto-Submit: George Engelbrecht <engeg@google.com>
Reviewed-by: Benjamin Shai <bshai@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>