api: Fix glob pattern for log files
Recent runs of CQ with failed emerge commands reveals that the package
name is available in the output proto, but the paths were not correctly
located. This CL edits the glob path to reflect the correct naming
scheme and emits a warning to build logs when no matching log files are
found.

BUG=b:204816060
TEST=unit
TEST=CQ forthcoming
TEST=ran gen_call_scripts after breaking net-wireless/floss locally.
Build log included the emitted warning before glob path was changed,
and now includes the log path in the output proto.

Change-Id: I0c29c4403c0f587ba9f5be2ef6de62555be935dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3336658
Tested-by: Lizzy Presland <zland@google.com>
Auto-Submit: Lizzy Presland <zland@google.com>
Reviewed-by: Alex Klein <saklein@chromium.org>