cbuildbot: deprecate cidb
The cidb isn't used for reporting anymore. We've replaced this with plx
based resources such as go/release- and go/cros-ci-builds-. There should
not be any more consumers of this DB.

This should reduce the amount of flakes related to talking to the
service while we turn down legacy.

BUG=b:247170885
TEST=units

Change-Id: I000825f33ee670410c389b71bdf88f89b1e1bad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3361315
Reviewed-by: Jack Neus <jackneus@google.com>
Reviewed-by: Benjamin Shai <bshai@google.com>
Tested-by: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>