gerrit/gob: switch http.client to urllib.request
The low level http library doesn't support *_proxy variables.  If we
want that, we'd have to implement it ourselves.  Switch to the higher
level urllib.request APIs which handles this for us.  We don't need
much here, just the request's response data & code, so we don't lose
anything by switching.

BUG=b:213649630
TEST=`gerrit account` works inside of `cros_sdk --proxy`

Change-Id: If1aa33abd44f39fb5f8867617070a2cf7fffe123
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3385060
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>