Make deploy_chrome pick up test data for the Finch smoke test
Add the file "variations_seed" to _COPY_PATHS_CHROME, which will cause
it to be picked up by the deploy_chrome script. This is a text file that
will be used by the ui.ChromeVariationsSmoke Tast test, as part of the
finch_smoke_test recipe.

The finch smoke test suite tests newly created finch configs against
different platforms. The finch config is the test input, and is captured
in the "variations_seed" file. The contents of this file are too large
to pass to a Tast test on the command line through a tast-var arg, so
instead the recipe will copy the file to the outdir so it can be picked
up by the deploy_chrome script which will move it to the DUT before the
Tast test runs.

BUG=b:202325793
TEST=add variations_seed file to the outdir, deploy_chrome, confirm file
is present on the device in /opt/google/chrome/.

Change-Id: Ib41e4fe123dde35cb95e038e7e7b0b99aedab2d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3425665
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>