Add support for loading Widevine CDM from it's own directory
The Widevine CDM is moving from the base install directory into it's
own directory named 'WidevineCdm' to allow it to include a manifest
indicating what is supported and in the future to allow it to be
component updated. This will match what is available on other
desktop platforms.

For now both the old way and the new way of including the Widevine
CDM are supported to allow for the transition. Once the change to
Chrome has been committed, the original location can be removed.

BUG=chromium:971433
TEST=successfully built and deployed image

Change-Id: I3ba8656e54964b991f22840168f9a3b42de185d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3426425
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Tested-by: John Rummell <jrummell@chromium.org>
Commit-Queue: John Rummell <jrummell@chromium.org>