SdkService: Add CreateBinhostCLs
When we build a new SDK, we need to change binhost to point at
the new packages. This adds a CreateBinHostCLs method to SdkService,
which creates CLs to point binhosts at a given SDK version.
The CLs are published to Gerrit, but not automatically committed.

BUG=b:196886631
TEST=Call the method from the build_toolchain recipe.

Change-Id: I5d9ffa4369c4eafc78e453e31eeb446fb58368f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3437304
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>