GeneratePayload: Gen miniOS payloads, if possible
Also some internal refactoring to support this change. Previously,
_Create returned a bool representing whether a payload was generated.
This omitted the critical information of _why_ a payload might not be
generated. I captured this information in the form of BaseException
subclasses, since exceptions are meant to represent breaks in the normal
codepath, and in this case the normal codepath for _Create is "creating"
a payload.

BUG=b:217556989
TEST=./run_tests.py

Change-Id: Ie94d5a046157df8dcd218ace019024bbdddf2604
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3456639
Reviewed-by: Benjamin Shai <bshai@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Auto-Submit: Greg Edelston <gredelston@google.com>
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>