image_test: Add hplip binaries to LinkageTest
In crrev.com/c/3227708 we accidentally caused hpcups to link against
libpython.  Add the hplip binaries to the LinkageTest list to make sure
this doesn't happen again.

This isn't a full fix for the general problem of linking against test
image libraries, but it ensures we catch binaries where we know this has
happened until the full fix is ready.

BUG=b:220149175
TEST=image_test chromiumos_base_image.bin without crrev.com/c/3425321

Change-Id: I76a6315a4e9f52b69098b3129ddaf12c0125e1d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3488684
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>