image_test: Check library resolution in LinkageTest
LinkageTest parses several binaries and all the ELF .so files in the
image root.  However, it doesn't check that the result of
lddtree.ParseELF actually resolved successfully, so the only thing it
was catching would have been invalid ELF files.

Instead, go through all the libraries listed in "needed" and make sure
each one resolved to a path.  When this is run against the base image,
anything that resolves to /usr/local will fail.  If there are other
undersirable paths, it should be straightforward to check for them
later.

BUG=b:220149175
TEST=image_test chromium_base_image.bin

Change-Id: I82d7373faef827c06b12121ad5e6395a2ebfa63f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3488683
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>