sysroot: Update GetUseFlags and GetEnv
In GetUseFlag, update to account for additional USE flags based on the
run_config.
Rename GetEnv to GetExtraEnv to more accurately reflect the nature of
the function. Update to specify any default environment variables. We
pass in the environment variables as extra_env to cros_build_lib.run
so that environment variables for the current process are used for the
new process and only the specified envvars in extra_env are used in
addition.

BUG=b:218521368
TEST=./run_tests service/sysroot_unittest.py

Change-Id: I6c8401ff0d1645ce57df2785916d8b0c32e9a5df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3504086
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Cindy Lin <xcl@google.com>
Tested-by: Cindy Lin <xcl@google.com>