cros_chrome_sdk: Download the autotest package unconditionally
The autotest package is inspected by Chrome's build during GN-gen to
build the list of test file dependencies. Since we're not downloading
that package when passing in --chroot, the GN files end up complaining
about the missing package when using --chroot.

So just download that package even when using the chroot to skip the
failure. There's as of yet not an easy/good way of determining if a
chroot is being used in Chrome's build files.

BUG=none
TEST=unittest

Change-Id: Id633b44729858878bdedbb0f23e9d4ed590cee69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3504328
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>