dependency_lib: Allow mapping masked package sources
In some cases the provided deps include masked packages that are
actually installed in the SDK. Allow the masked packages to have
their sources mapped to prevent unnecessary errors.

The situation that surfaced the issue was running
`cros workon -b tatl start .`. The depgraph for tatl includes
python 2.7, which is masked for the sysroot, but is only included
as a bdepend. Having cros workon only generate the sysroot depgraph
provided an alternative solution, but in order to allow devs to
workon packages installed to the SDK, this solution was preferred.

BUG=b:214874483
TEST=manual: cros workon -b tatl start .

Change-Id: Ie91e4261fd2850a34f55b1206289e154077a33f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3508441
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>