cros-ide: Bug fixes for default board setting
We have been storing the default board config per workspace. It means it
updates .vscode in user's workspace and could pollute the team's config
committed to git. This patch updates the config to be stored globally,
addressing the issue.
This patch also addresses the issue that the function
selectAndUpdateTargetBoard doesn't return in case the message dialog is
dismissed due to timeout.

BUG=b:223087575
TEST=manually confirmed that the config is stored to User instead of
Workspace.

Change-Id: I9b501950115f991d898916f6b4e13e41bc57ce4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3521154
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>