cros-ide: Show error if compilation database not generated
With this patch, when compilation database is not generated, the IDE
will show an error message suggesting the user to update their ebuild
file to generate it. If user clicks the Dismiss button, the error
message will not reappear for the current instance.

BUG=b:214323485
TEST=Error message is shown for spaced package. Once it's dismissed, it
doesn't reappear.
TEST=Error message is not shown for cros-disks package.

Change-Id: I575e875c253a8752cf9c38cc8fe5ee4469d4d2f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3526806
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>