cros-ide: do not throw errors in exec, return them instead.
We change the return type of exec from Promise<ExecResult> to
Promise<ExecResult|Error>. Accessing stdout and other properties
of ExecResult without checks for Error will result in compilation error.

We do not change the behavior of the callers - we explicitly throw
errors that were allowed to propagate earlier.

Note, that errors will be silently ignored, if the result of a call to
exec() is unused.

BUG=b:223488009
TEST=npm run test

Change-Id: I26a616f9c04e733a0c91b89b3b60777ffdd43532
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3562095
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>