api/toolchain: Make filters optional for linting
This CL contains the changes generated by compile_build_api_proto after
adding a new field to the LinterRequest proto to make filtering
lints by whether or not they concern modified lines optional.

This CL also implements the change in the endpoint to actually use it,
and also updates the associated call_template.

This is to support using these endpoints to query for the list of
current lints in platform2.

BUG=b:229121199
TEST=./compile_build_api_proto

Change-Id: Ic309ea718855e0da19d3c688ee8007ba29635a9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3585344
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>