toolchain: rm build caches in linting endpoints
This CL implements removing the portage temp files from previous calls to
emerge for packages being linted. This is helpful behavior for
because linting artifiacts are not produced if emerge skips the package
due to there being no changes since the last build. This doesn't matter
on builders, but makes it easier to use the endpoint locally.

BUG=b:229143726, b:187790543
TEST=tested with call_scripts

Change-Id: Ib7a3b4c7481f968a3e9dec708fe2c7b762ddda41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3586244
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>