toolchain: move linting logic from api to service
This CL begins a refactor of the toolchain controller in the Chromite
API to move implementation details into chromite/service as is the
standard design of other services. The specific scope of this CL is
limited to the EmergeWithLinting endpoint and its supporting functions.

We also take the opportunity to restructure linting functions into a
new class for readability.

Moving other endpoints over to the new service/toolchain.py location is
also desirable though saved for a future CL.

BUG=b:229665884
TEST=api/call_scripts/toolchain__emerge_with_linting

Change-Id: I26ece117ee36f05bfc256f31927e4ed84ab8d1f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3603370
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>