Add external config flag to deploy_chrome
Adds the --use-external-config flag to deploy_chrome.py, and
by extension, cros_test.py. This gets auto-added when
--public-image is passed in.

This is necessary in order to deploy a stripped Chrome binary
to devices that are using -full (as opposed to -release) images.
Without specifying to use the external config, the host can end
up trying to download from the internal GCS bucket when
stripping the binary.

This will be used in Chromium to strip Chrome on Kevin devices
for certain tests to work around Breakpad running out of memory
when generating a minidump.

Bug: chromium:1313191
Change-Id: Ibe023dbcb886b97971799c7bfec1a9d4829700fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3608462
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Tested-by: Brian Sheedy <bsheedy@chromium.org>