cros-ide: adding 'upstart' language.
We do not have an accurate way of telling if a file is an upstart
script, because .conf is a generic extension used in many file types,
and we cannot use the first line to tell it either. So for now, users
will have to select the language manually.

In follow-up changes we will add syntax highlighting and configure
editor features, such as comment toggling.

BUG=b:230052562
TEST=verified that 'Upstart' can be selected as a language in the status
bar

Change-Id: Ic919e77b5b6b02f2676924b7a13333515fe7857c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3609872
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>