contrib: zephyr_downstream: add support for limit and stop-at
Add two flags to the script: limit, to only tag the most recent changes,
and stop-at, to stop tagging changes when hitting a specific one. These
may be helpful to upstream a bunch of changes until a problematic one,
so that they can go through while the others are being worked on.

BUG=None
TEST=actually used the flags

Change-Id: I931a0c40fbd67ab13309c8e65ffebd8ee8eb1c30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3613807
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Benjamin Shai <bshai@google.com>