parallel: pass the PID as a string to _DumpDebugPid
We noticed some string encoding errors in the stdout of a failed build
when cros_run_test went to shut down a VM that was hanging. This change
should make it so that the PID is always a string.

BUG=b:231141194
TEST=./run_tests

Change-Id: I387c9f20d6a50e59680c21cb82d3c289b51d29ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3620659
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Joseph Sussman <josephsussman@google.com>
Tested-by: Joseph Sussman <josephsussman@google.com>
Auto-Submit: Joseph Sussman <josephsussman@google.com>