remote_access: add type info for basic pathlib support
Some of these APIs already support pathlib as-is, so document it in
with typing info.  The tempdir member needs more work, so coerce it
to a string for now so we can still pass it a Path object as we add
support to other APIs that use remote_access.

BUG=None
TEST=`./run_tests` passes

Change-Id: I9dc698b1d60fcdd0a12618d87bf4453eb36c1b9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3625734
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Lizzy Presland <zland@google.com>
Commit-Queue: Lizzy Presland <zland@google.com>