api: Stop populating repeated PackageInfo field
In preparation for deprecating and reserving the old repeated
PackageInfo field which only contained the names of the failed packages,
we want to stop populating and referencing the field. Recipes no longer
references this field for retrieving and displaying the names of failed
packages, so it is safe to stop populating the field in response protos.

BUG=b:204816060
TEST=./run_tests

Change-Id: Id73ee535d07f662fbf8582f342258579ed4284c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3631025
Auto-Submit: Lizzy Presland <zland@google.com>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Tested-by: Lizzy Presland <zland@google.com>
Commit-Queue: Lizzy Presland <zland@google.com>