cros lint: switch parallel APIs
Avoiding the Queue & Value structures seems to speed cros lint up a
bit.  It means we have to compute the entire task list up front, but
that is almost always faster unless we're trying to lint thousands of
files.  Which we pretty much never do.

This also simplifies the code a bit.

BUG=None
TEST=CQ passes

Change-Id: Ibe68103098293197d4b5bb670c7ecd0cf43fed74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3643401
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gilberto Contreras <gcontreras@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>