cros lint: generalize "linter" to "tool"
The code in here shares a bit with cros format, and it uses "tool"
rather than "formatter" in many places.  Changing the naming in the
main part of this file makes it a bit more obvious that the two have
a bit of overlap and hopefully make code sharing easier.

BUG=None
TEST=`cros lint` on files still works

Change-Id: I0d0c528d84020fc549364de4dcc4dc60410792cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3643402
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Ram Chandrasekar <rchandrasekar@google.com>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>