[Lacros] Add variation_seed file to deploy_chrome
Make deploy_chrome pick up test data for the Finch smoke test

Add the file "variations_seed" to _COPY_PATHS_LACROS, which will cause
it to be picked up by the deploy_chrome script. This is a text file that
will be used by the lacros.VariationSmoke Tast test, as part of the
finch_smoke_test recipe.

deploy_chrome script will copy the seed file to lacros dir in the DUT
and it can be used in the tast test.

BUG=chromium:1305074
TEST=verified with command "./third_party/chromite/bin/deploy_chrome --build-dir=out_vm_lacros/Release --device=DUT --nostrip --lacros"

Change-Id: If3d1c1882c415c8b042049f590edead1387bcfce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3646851
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Tested-by: Jieting Yang <yjt@google.com>
Commit-Queue: Jieting Yang <yjt@google.com>